+++ src/sdk/PnP.Core/Model/SharePoint/Pages/Public/Enums/PageHeaderType.cs
@@ -21,7 +21,7 @@
         Custom = 2,
 
         /// <summary>
-        /// All Headersettings are in the PageTitle WebPart. PageTitle-WP in first section as OneColumnFullWith (Message ID: MC791596 / Roadmap ID: 386904)
+        /// All Headersettings are in the PageTitle WebPart. PageTitle-WP in a section as OneColumnFullWith (Message ID: MC791596 / Roadmap ID: 386904)
         /// </summary>
         PageTitleWebPart = 3,
     }